refactor(FullscreenNav): remove the interface to index file
