--- src/components/navigation/FullscreenNav/interfaces/FullscreenNav.FullscreenMenu.interface.ts
@@ -1,8 +1,0 @@
-import { INavigation } from "./FullscreenNav.Navigation.interface";
-
-export interface IFullscreenMenuProps {
-  navigation: INavigation;
-  logoutPath: string;
-  logoutTitle: string;
-  onClose: () => void;
-}
