--- src/components/navigation/FullscreenNav/interfaces/FullscreenNav.Link.interface.ts
@@ -1,6 +1,0 @@
-export interface ILink {
-  id: string;
-  label: string;
-  icon: React.ReactNode;
-  path: string;
-}
