--- src/components/navigation/FullscreenNav/interfaces/FullscreenNav.Navigation.interface.ts
@@ -1,6 +1,0 @@
-import { ISection } from "./FullscreenNav.Section.interface";
-
-export interface INavigation {
-  title: string;
-  sections: { [key: string]: ISection };
-}
