--- src/components/navigation/FullscreenNav/interfaces/FullscreenNav.Section.interface.ts
@@ -1,6 +1,0 @@
-import { ILink } from "./FullscreenNav.Link.interface";
-
-export interface ISection {
-  name: string;
-  links: { [key: string]: ILink };
-}
