--- src/components/navigation/FullscreenNav/interfaces/FullscreenNav.interface.ts
@@ -1,8 +1,0 @@
-import { INavigation } from "./FullscreenNav.Navigation.interface";
-
-export interface IFullscreenNavProps {
-  portalId: string;
-  navigation: INavigation;
-  logoutPath: string;
-  logoutTitle: string;
-}
