--- src/components/navigation/FullscreenNav/interfaces/Fullscreennav.MultiSections.interface.ts
@@ -1,5 +1,0 @@
-import { INavigation } from "./FullscreenNav.Navigation.interface";
-
-export interface IMenuSectionsProps {
-  navigation: INavigation;
-}
