+++ src/components/navigation/FullscreenNav/stories/FullscreenNav.WithoutSections.stories.tsx
@@ -10,9 +10,8 @@ import {
   MdAccountBalanceWallet,
 } from "react-icons/md";
 
-import { FullscreenNav } from "..";
+import { FullscreenNav, IFullscreenNavProps } from "..";
 import { parameters, portalId, navigation, logoutPath } from "./props";
-import { IFullscreenNavProps } from "../interfaces/FullscreenNav.interface";
 
 const story = {
   title: "navigation/FullscreenNav",
